3.2.81

+++ src/XMusic/cpp/bkgdlg.cpp
@@ -109,7 +109,7 @@ void CBkgView::_onPaintRow(CPainter& painter, tagLVRow& lvRow)
 
             auto d = (abs(g_crText.red()-g_crTheme.red()) + abs(g_crText.green()-g_crTheme.green())
                 + abs(g_crText.blue()-g_crTheme.blue()))/3;
-            int nAlpha = 255-(255*pow((float)d/255,0.5));
+            int nAlpha = 255-(255*pow(d/255.0,0.5));
             nAlpha = MAX(nAlpha, 85);
             QColor cr = g_crText;
             cr.setAlpha(nAlpha);
