+++ src/XMusic/cpp/listview.cpp
@@ -94,7 +94,7 @@ void CListView::_onPaint(CPainter& painter, int cx, int cy)
             {
                 auto d = (abs(g_crText.red()-g_crTheme.red()) + abs(g_crText.green()-g_crTheme.green())
                     + abs(g_crText.blue()-g_crTheme.blue()))/3;
-                int nAlpha = 255-(255*pow((float)d/255,0.1));
+                int nAlpha = 255-(255*pow(d/255.0,0.1));
                 nAlpha = MAX(nAlpha, 32);
                 QColor cr = g_crText;
                 cr.setAlpha(nAlpha);
@@ -160,7 +160,7 @@ void CListView::_paintRow(CPainter& painter, const tagLVRow& lvRow, const tagRow
     {
         auto d = (abs(g_crText.red()-g_crTheme.red()) + abs(g_crText.green()-g_crTheme.green())
             + abs(g_crText.blue()-g_crTheme.blue()))/3;
-        int nAlpha = 255-(255*pow((float)d/255,0.5));
+        int nAlpha = 255-(255*pow(d/255.0,0.5));
         nAlpha = MAX(nAlpha, 85);
         QColor cr = g_crText;
         cr.setAlpha(nAlpha);
